not inf > 0.6 on SC, interesting result
There was not enough penalty on the high cost side of the matrix so red
elements were often being selected. This blocks that behaviour. It makes
the contig/symmetry run better but the basic SUM is still just way too
good. Close to giving up. Can still try pre-processing C and using the
standard SUM approach to avoid the current approaches