First demonstrable improvement over the summation matrix
Displays mean diff and shift on cost matrix, new "heuristic" evaluation
which compares indices to the nearest predictions to make it alignment
invariant, this will be our primary measurement of accuracy from now on.
The contig cost function takes a vector of all elements in the square,
sorts them, then takes the mean positive first order difference
normalised by the number of elements in the square, this is them
multiplied with the summation matrix normalized by W (not W^2 which
regularises the placement). Average shift is displayed at the end to
give insight if we are predicting too late or early on average which
will allow us to align our error variable with the human error variable.
I also observed on the data that the human indices seem to be more
inaccurate towards the end of the shows, perhaps when the human is
getting tired/bored