Decent working prototype of the contiguity cost function!
This cost function gets a concatenated line vector from the square of
all adjacent diagonals, adjacent value pairs are compared and blues get
score + 1, reds get score -1. The score is normalized by the number of
elements in square. While the accuracy isn't brilliant, it looks
extremely good and seems to perform robustly, even working on the
nightmare ASOT 462! In conjunction with the others it might improve
their robustness.