Config settings in separate file, tweaked the contig map, performance now comparable to the main method
I also expect it to be more robust some of the time. It makes a
catastrophic error on the dreaded ASOT 462 but I wouldn't be able to
index that one myself by looking at the image, it's more of a mystery
how the summation matrix gets it right. This seems to be doing the exact
job I would do looking at the cost matrix and labelling them myself