More breakthroughs
Amazingly only just realised that the summation cost method was only
counting disincentives i.e. the zeros were not counted. I now transform
it onto a [-1,1] interval and it looks much more similar to the other
cost functions. I have no idea why it wasn't always like this although
as a caveat it performs slightly worse but I bet its more robust now.
Cost matrix regularization exponents have been factored out and also the
gauss win is now counted as a cost matrix which is good for me to test
the naive algorithm and it can be used to regularise the other matrices.
The other normalization types 1,2 I now realise were compensating for
the flawed design of the old cost matrix