Massive overhaul
*Heuristic metric now returns the median of the closest matches, which
is far more descriptive
*moved thresholds into the contig/symmetry as semi perm parameters
*I now detect the skew of the cosine matrix and shift the mean to 0.5,
this makes all methods work more consistently and is especially
important for the contig/sym method. Cant believe I never did it before.
I can't explain the cause of the skew (it is normally distributed).
Either our extraction or the underlying show
* removed some old parameters
* overhauled the scaling and merging of cost functions, cosine matrix
now scaled to [-1.1] and so are all the cost matrices, they are all
simply added together with no further normalization
* overhauled all three cost functions from scratch
* sum cost function is now rewritten in a slow way (no DP) but does some
clever normalization. It might be possible to modify the old DP routine
to produce the same.
* all three cost functions now have incentive and disincentive as first
class citizens. funnily enough the concept doesn't seem to help but does
explain a lot about the weird scaling that was required before to get it
to work (when it was a disincentive only function).
* generally on the incentive, I multiply by a gaussian window and not no
disincentive
* it's got to the stage now where I am convinced in almost every case we
produce better track indices than the humans do. The error metric is
making us look worse the more accurate we actually get! I realised that
on the other paper it worked well because by normalizing by w instead of
w^2 we were accidently adding a kind of gaussian blur or error variable
to the cost matrix and that was the reason that the results were
slightly better

*