Code cleanup, results parser, contig now normalizes by progression
Running an experiment is too slow, contig takes about 45 seconds and
symmetry slightly less, this is crazy. First idea was to mix them
together into one function as they are compatible to give a 50% speed
up. Just realised that there is a dynamic programming recursion for
calculating symmetry and an an obvious one for contig which will save an
order of magnitude time complexity. The slow results are looking
brilliant so we are all set with the method at least