various improvements
Subplots
Slight perf increase on contig (might not be quite as good, not using
mean).
Rewritten the gaussian/first derivative code to be cleaner
The main loop in get_cosine.. is much faster, I subsample to a fixed
number of elements and don't do abs() on the whole complex fft vector
Convolution is way faster because the audio vector is smaller and the
kernel is smaller
This might be enough to eliminate some of the memory errors when doing
the genetic algorithm search over 8 instances, will test on lower tile
sizes though. The contig and symmetry cost matrices start to come out
slower than the DP search when tile size <= 6
Simplified code in resample_vector